Move some non-duplicated FS related functions to preload
